%d for int64_t changed to %lld
When using /xp to get the average, the output for total run time would show minutes in the seconds. Looking into it, we were using %d to display int64_t. After correcting this location Sic noted the StartTime.systicks also having the same issue for the output, so corrected that as well.